+++ lib/fisk8viewer/parsers.rb
@@ -11,6 +11,15 @@ module Fisk8Viewer
   end
 end
 
+## require all rb files under ./parsers
+Dir[File.expand_path('../parsers', __FILE__) << '/*.rb'].each do |file|
+  require file
+end
+
+#ActiveSupport::Dependencies.autoload_paths << "lib/fisk8viewer/parsers"
+=begin
 require 'fisk8viewer/parsers/isu_generic'
 require 'fisk8viewer/parsers/isu_generic_mdy'
 require 'fisk8viewer/parsers/wtt_2017'
+require 'fisk8viewer/parsers/finlandia'
+=end