+++ lib/fisk8viewer/parsers/wtt_2017.rb
@@ -1,5 +1,3 @@
-
-
 module Fisk8Viewer
   module Parsers
     class WTT_2017 < ISU_Generic